refactor(sponsor-reports): address PR #997 convention review (smarcet)
- index.js: import DEFAULT_PER_PAGE from constants (drop local DEFAULT_PAGE_SIZE dup)
- index.js: mapDispatchToProps to object shorthand via new clearPurchaseDetailsValidation
  action creator (removes the raw dispatch prop)
- OrdersTable.js: remove dead getOrderRowId export + its test (page maps id inline)
- usePrint: move to src/hooks/ per the codebase custom-hook convention
- section-csv-query.js: align defense-in-depth comment with the sponsor_id/page_id guard
