+++ src/actions/sponsor-reports-actions.js
@@ -61,6 +61,13 @@ export const getPurchaseDetailsReport =
       .finally(() => dispatch(stopLoading()));
   };
 
+// Clears the Purchase Details validation toast (dispatched from the Snackbar
+// onClose). A plain action creator lets the page bind it via the object form of
+// mapDispatchToProps instead of receiving raw dispatch.
+export const clearPurchaseDetailsValidation = () => ({
+  type: PURCHASE_DETAILS_VALIDATION_CLEAR
+});
+
 export const getPurchaseDetailsLinesReport =
   (query = {}) =>
   async (dispatch, getState) => {
